requirements.txt
torch
transformers
accelerate
datasets
peft
trl
bitsandbytes
sentencepiece
scikit-learn
pandas
numpy
pyyaml
jupyter
